ENG-3517: Fix CR
